update structures on Raspberry code
